+++ app/Filament/Pages/Tenancy/EditCompanySettings.php
@@ -185,118 +185,6 @@ class EditCompanySettings extends EditTenantProfile
                     ->collapsed()
                     ->columns(2),
 
-                Section::make('Scheduled Exports')
-                    ->description('Configure recurring automated Tally XML exports that are emailed to specified recipients on a schedule.')
-                    ->schema([
-                        Repeater::make('scheduledTallyExports')
-                            ->relationship()
-                            ->label('')
-                            ->schema([
-                                Select::make('frequency')
-                                    ->options(ExportFrequency::class)
-                                    ->required()
-                                    ->live()
-                                    ->default(ExportFrequency::Weekly),
-                                Select::make('day_of_week')
-                                    ->label('Day of Week')
-                                    ->options([
-                                        0 => 'Sunday',
-                                        1 => 'Monday',
-                                        2 => 'Tuesday',
-                                        3 => 'Wednesday',
-                                        4 => 'Thursday',
-                                        5 => 'Friday',
-                                        6 => 'Saturday',
-                                    ])
-                                    ->default(1)
-                                    ->visible(fn (callable $get): bool => $get('frequency') === ExportFrequency::Weekly->value)
-                                    ->required(fn (callable $get): bool => $get('frequency') === ExportFrequency::Weekly->value),
-                                Select::make('day_of_month')
-                                    ->label('Day of Month')
-                                    ->options(array_combine(range(1, 28), range(1, 28)))
-                                    ->default(1)
-                                    ->visible(fn (callable $get): bool => $get('frequency') === ExportFrequency::Monthly->value)
-                                    ->required(fn (callable $get): bool => $get('frequency') === ExportFrequency::Monthly->value)
-                                    ->helperText('Capped at 28 to avoid end-of-month issues.'),
-                                TimePicker::make('time_of_day')
-                                    ->label('Time')
-                                    ->default('10:00')
-                                    ->seconds(false)
-                                    ->required(),
-                                Select::make('timezone')
-                                    ->options([
-                                        'Asia/Kolkata' => 'IST (Asia/Kolkata)',
-                                        'UTC' => 'UTC',
-                                        'America/New_York' => 'EST (America/New_York)',
-                                        'Europe/London' => 'GMT (Europe/London)',
-                                        'Asia/Dubai' => 'GST (Asia/Dubai)',
-                                        'Asia/Singapore' => 'SGT (Asia/Singapore)',
-                                    ])
-                                    ->default('Asia/Kolkata')
-                                    ->required(),
-                                Select::make('date_range_window')
-                                    ->label('Date Range')
-                                    ->options(DateRangeWindow::class)
-                                    ->required()
-                                    ->default(DateRangeWindow::Previous7Days),
-                                Select::make('statement_type')
-                                    ->label('Transaction Type')
-                                    ->options([
-                                        '' => 'All Transactions',
-                                        ...array_column(
-                                            array_map(
-                                                fn (StatementType $type) => ['value' => $type->value, 'label' => $type->getLabel()],
-                                                StatementType::cases()
-                                            ),
-                                            'label',
-                                            'value'
-                                        ),
-                                    ])
-                                    ->placeholder('All Transactions')
-                                    ->helperText('Filter by Bank, Credit Card, or Invoice transactions. Leave blank for all.'),
-                                TagsInput::make('recipient_emails')
-                                    ->label('Recipient Emails')
-                                    ->placeholder('Enter email addresses')
-                                    ->required()
-                                    ->nestedRecursiveRules(['email'])
-                                    ->columnSpanFull(),
-                                Toggle::make('is_active')
-                                    ->label('Active')
-                                    ->default(true)
-                                    ->columnSpanFull(),
-                                Placeholder::make('last_run_status')
-                                    ->label('Status')
-                                    ->content(fn (?ScheduledTallyExport $record): string => $record?->last_run_status ?? 'Never run')
-                                    ->visible(fn (?ScheduledTallyExport $record): bool => $record !== null),
-                                Placeholder::make('last_run_at')
-                                    ->label('Last Run')
-                                    ->content(fn (?ScheduledTallyExport $record): string => $record?->last_run_at?->diffForHumans() ?? 'N/A')
-                                    ->visible(fn (?ScheduledTallyExport $record): bool => $record !== null),
-                                Placeholder::make('last_run_message')
-                                    ->label('Message')
-                                    ->content(fn (?ScheduledTallyExport $record): string => $record?->last_run_message ?? '-')
-                                    ->visible(fn (?ScheduledTallyExport $record): bool => (bool) $record?->last_run_message)
-                                    ->columnSpanFull(),
-                            ])
-                            ->columns(2)
-                            ->columns(2)
-                            ->defaultItems(0)
-                            ->addActionLabel('Add Schedule')
-                            ->mutateRelationshipDataBeforeCreateUsing(function (array $data): array {
-                                $data['created_by'] = Auth::id();
-                                $data['statement_type'] = $data['statement_type'] ?: null;
-
-                                return $data;
-                            })
-                            ->mutateRelationshipDataBeforeSaveUsing(function (array $data): array {
-                                $data['statement_type'] = $data['statement_type'] ?: null;
-
-                                return $data;
-                            }),
-                    ])
-                    ->collapsible()
-                    ->collapsed(),
-
                 Section::make('Integrations')
                     ->schema(fn () => $this->getIntegrationsSchema()),
             ]);
@@ -401,54 +289,6 @@ class EditCompanySettings extends EditTenantProfile
 
                     Notification::make()
                         ->title('Zoho Invoice disconnected.')
-                        ->success()
-                        ->send();
-                }),
-            Action::make('testExportNow')
-                ->label('Test Export Now')
-                ->color('warning')
-                ->icon('heroicon-o-paper-airplane')
-                ->requiresConfirmation()
-                ->modalHeading('Test Export Now')
-                ->modalDescription('Select a scheduled export to trigger a test run immediately. The export will be sent to the configured recipients.')
-                ->schema([
-                    Select::make('schedule_id')
-                        ->label('Schedule')
-                        ->options(function (): array {
-                            /** @var Company|null $company */
-                            $company = Filament::getTenant();
-
-                            if (! $company) {
-                                return [];
-                            }
-
-                            return $company->scheduledTallyExports()
-                                ->where('is_active', true)
-                                ->get()
-                                ->mapWithKeys(fn (ScheduledTallyExport $s) => [
-                                    $s->id => $s->schedule_description.' — '.implode(', ', $s->recipient_emails),
-                                ])
-                                ->all();
-                        })
-                        ->required(),
-                ])
-                ->visible(function (): bool {
-                    /** @var Company|null $company */
-                    $company = Filament::getTenant();
-
-                    return $company?->scheduledTallyExports()->where('is_active', true)->exists() ?? false;
-                })
-                ->action(function (array $data): void {
-                    $schedule = ScheduledTallyExport::find($data['schedule_id']);
-
-                    if (! $schedule) {
-                        return;
-                    }
-
-                    SendScheduledTallyExport::dispatch($schedule);
-
-                    Notification::make()
-                        ->title('Test export dispatched — check your email shortly.')
                         ->success()
                         ->send();
                 }),
