+++ resources/views/filament/pages/scheduled-exports.blade.php
@@ -1,0 +1,3 @@
+<x-filament-panels::page>
+    {{ $this->content }}
+</x-filament-panels::page>
